replenishment_v2/requirements.txt
# Core
torch>=2.0.0
numpy>=1.21.0
pandas>=1.4.0
pyarrow>=8.0.0

# RL & Training
gymnasium>=0.28.0
tensorboard>=2.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Optional: Distributed Training
# torch-distributed

# Development
black>=23.0.0
pytest>=7.0.0
